TechManager Tech Tree Support
Add a sub-tree, or "branch" of the stock tree that contains Impossible
Innovations parts. 2 parts, the aluminum strut and the dead weight are
located in the stock tree. Powered by TechManager.

Removed some typos, played with more price balancing.

Discovered that not every part cfg needs to be named 'part.cfg'! So I
renamed every single one to make it easier for me to edit.

New models for struts, radial battery, and 2 tanks!

Radial battery is now scalable.