Finalizing the Tree
Made a few corrections in the tree cfg, and also got a non-bugged
version of TechManager for a non-buggy playing experience!

I apparently missed some parts in my "set all node sizes to 1" sweep, so
I got them in this commit.

If nothing goes wrong when I test this out, this commit will be the next
release!