I Fixed Some Typos
Fixed a typo in the aluminum strut.cfg that prevented it from showing up
in the tech tree.

The research node title "Alternate Methods of Nuclear Propulsion" was
too long to fit in its slot in the R&D screen, so I renamed it to "Dawn
of Fusion." Definitely more dramatic, I say.